requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
torch
torchvision
efficientnet-pytorch
numpy
python-decouple
aiofiles
scikit-learn
opencv-python-headless
